+++ packages/create-mastra/CHANGELOG.md
@@ -1,5 +1,13 @@
 # create-mastra
 
+## 1.16.1-alpha.5
+
+## 1.16.1-alpha.4
+
+## 1.16.1-alpha.3
+
+## 1.16.1-alpha.2
+
 ## 1.16.1-alpha.1
 
 ## 1.16.1-alpha.0